Revert disabling atime for folders (#13)
We disabled tracking the atime of folders for testing the functionality
of the file expiry emails. This was because previous versions of this
library had been updating the atime of folders while collecting data and
so the atime info is not useful to us right now.

We should change to using this version of the library after another
month and a half have passed at least.