Added option to use or disuse atime for directories (#14)
Added optional flag to use or disuse the atime for folders when checking
expiry

We are temporarily disabling the use of atime for folders. This is
because previous testing in this project updated the atime of all
folders in the /var/lib/clusters/users directory, and so it is not
valuable to check the atime or those folders anymore.