Update scripts to generate two files for creator info - seperated by expiry and deletion (#20)
Currently we just generate a single overarching file for creator info,
then later check whether such files are expiring or should be deleted.

Later in the file expiry / deletion pipeline, we will have to send
emails based on if files are expired or should be deleted. Then for MVP,
an admin should manually delete the files to be deleted. This change
would streamline the process by just separating that info into different
files now, rather then having to parse and separate this information
later.